Major efficiencly improvements
Performs collect before transforming RDD[ping] to DataFrame. Without
this operation, the parallelism is reduced to the number of partitions
used when writing to parquet. I shit you not.

Moves the testpilot experiment filter to before the transformations.
There's no need to transform every ping, just the ones for this
experiment.